Adds historical weather queries with example; reduces code redundancy; adds properties
This is a large commit which basically updates the code to be more
useful for my purposes.  I've added a method to query historical
weather, although the API itself is a bit fragile when I've tested it.
To the example code, I've added a gesture based transition to show the
historical query (yesterday's weather).  Some amount of code redundancy
was eliminated through common methods and properties.  I've also
updated the code style to match other parts of my project, which is
mostly based on the Google Objective-C style.